pyproject.toml
[tool.poetry]
name = "customer_segmentation"
version = "0.1.0"
description = ""
authors = ["Khuyen <khuyentran1476@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
numpy = ">=1.14.3,<1.21.0"
matplotlib = "^3.5.1"
scikit-learn = "^1.0.2"
hydra-core = "^1.1.1"
dvc = {extras = ["gdrive"], version = "^2.9.3"}
yellowbrick = "^1.3.post1"
pandas = "<1.3.5"
bentoml = "1.0.0a4"
pydantic = "^1.9.0"
streamlit = "^1.5.1"
bentoctl = "1.0.0a4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''

